refactor: move Props types to component file headers
- Move CircleOfFifthsProps to CircleOfFifths.tsx
- Move CircleSegmentProps to CircleSegment.tsx
- Move KeyAreaProps to KeyArea.tsx
- Remove obsolete types/props.ts file
- Update useKeyArea hook imports and type references
- Fix type alias conflicts with component names
- Update documentation to reflect new Props location
- Update development agreement for Props type organization
