+++ docs/03.developmentAgreement.md
@@ -29,7 +29,7 @@
 - **TypeScript**:
   - `any` 型の安易な使用は原則として禁止する。型推論が困難な場合に限り、その理由をコメントで明記した上で使用を検討する。
   - コンポーネントの `Props` や関数の引数・返り値には、明確な型定義を付与すること。
-  - 型定義は、コンポーネント固有のものはコンポーネントディレクトリ内の `types` に、グローバルなものは `src/types` に配置する。
+  - 型定義は、コンポーネント固有のProps型は各コンポーネントファイルの冒頭に定義し、グローバルなものは `src/types` に配置する。
 
 ### 2.5. コメント
